+++ packages/xchain-evm/rollup.config.js
@@ -1,7 +1,6 @@
 import commonjs from '@rollup/plugin-commonjs'
 import json from '@rollup/plugin-json'
 import resolve from '@rollup/plugin-node-resolve'
-import external from 'rollup-plugin-peer-deps-external'
 import typescript from 'rollup-plugin-typescript2'
 
 import pkg from './package.json'
@@ -24,7 +23,6 @@ export default {
   ],
   plugins: [
     // ignore(["@ethersproject/providers", "@ethersproject/abstract-provider", "@ethersproject/strings"]),
-    external(),
     resolve({ preferBuiltins: true, browser: true }),
     typescript({
       tsconfig: './tsconfig.json',
@@ -33,5 +31,5 @@ export default {
     commonjs(),
     json(),
   ],
-  external: ['buffer', 'http', 'https', 'url', 'stream', 'string_decoder'],
+  external: Object.keys(pkg.dependencies || {}).concat(Object.keys(pkg.peerDependencies || {})),
 }