+++ packages/xchain-thorchain/rollup.config.js
@@ -1,7 +1,6 @@
 import commonjs from '@rollup/plugin-commonjs'
 import json from '@rollup/plugin-json'
 import resolve from '@rollup/plugin-node-resolve'
-import external from 'rollup-plugin-peer-deps-external'
 import typescript from 'rollup-plugin-typescript2'
 
 import pkg from './package.json'
@@ -28,14 +27,13 @@ export default {
   },
   plugins: [
     json({}),
-    external(),
     resolve({ preferBuiltins: true, browser: true }),
     typescript({
       exclude: '__tests__/**',
     }),
     commonjs({
-      exclude: '**/*.json',
+      esmExternals: true,
     }),
   ],
-  external: ['readable-stream', 'axios', 'buffer', 'crypto', 'stream', 'string_decoder'],
+  external: Object.keys(pkg.dependencies || {}).concat(Object.keys(pkg.peerDependencies || {})),
 }