+++ examples/estimate-swap/CHANGELOG.md
@@ -1,5 +1,18 @@
 # xchainjs-estimate-swap
 
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies [b93add9]
+  - @xchainjs/xchain-thorchain-query@0.7.10
+  - @xchainjs/xchain-midgard-query@0.1.15
+  - @xchainjs/xchain-thornode@0.3.13
+  - @xchainjs/xchain-midgard@0.5.5
+  - @xchainjs/xchain-client@0.16.4
+  - @xchainjs/xchain-crypto@0.3.2
+  - @xchainjs/xchain-util@0.13.4
+
 ## 1.0.2
 
 ### Patch Changes